refactor(docs): move components to theme
