--- website/src/components/PropsTable.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-type PropsTableProps = {
-	from: typeof React.Component;
-}
-
-export const PropsTable = ({ from: component }: PropsTableProps): JSX.Element => {
-	console.log(component);
-	return (
-		<table />
-	);
-};
