--- website/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from './Color';
-export * from './PropsTable';
-export * from './Seagull';
